Remove NavList
